pyproject.toml
[project]
name = "email-manager"
version = "0.1.0"
description = "Serverless Gmail email categorizer with AI-powered triage and Slack notifications"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "anthropic>=0.80.0,<1.0.0",
    "google-api-python-client>=2.150.0,<3.0.0",
    "google-auth>=2.35.0,<3.0.0",
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "pydantic>=2.10.0,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "moto[ssm]>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
